requirements.txt
streamlit
yt-dlp
python-dotenv
openai-whisper
scikit-learn
